data validation

+++ api/controllers/BookController.cs
@@ -13,92 +13,33 @@ namespace api.controllers {
         private readonly IBookRepository _bookRepo;
         private readonly ICustomerRepository _customerRepo;
 
-        public BookController(/*ApplacationDBContext context, */IBookRepository bookRepo, ICustomerRepository customerRepo) {
-            // _context = context;
+        public BookController(IBookRepository bookRepo, ICustomerRepository customerRepo) {
+;
             _bookRepo = bookRepo;
             _customerRepo = customerRepo;
         }
         
-        // [HttpGet]
-        //  public async Task<IActionResult> GetAll() {
-        //    var books = _context.Book.ToList();
 
-        //    if(books == null) {
-        //         return BadRequest("adsadsadsadsad");
-        //    }
-        //    return Ok(books);
-        // }
-
-        // [HttpGet("{id}")]
-        // public async Task<IActionResult> GetById(int id) {
-        //     var books = _context.Book.Find(id);
-
-        //     if(books == null) {
-        //         return BadRequest("Not Found Book ID THIS");
-        //    }
-        //    return Ok(books);
-        // }
-
-        // [HttpPost]
-        // public async Task<IActionResult> Create([FromBody] Book book) {
-        //     if (book == null) {
-        //         return BadRequest("Invalid book data 3999.");
-        //     }
-        //     try {
-        //         _context.Book.Add(book);
-        //         await _context.SaveChangesAsync();
-        //         return Ok(book);
-        //     }
-        //     catch (Exception ex) {
-        //         return StatusCode(500, $"Internal server error: {ex.Message}");
-        //     }
-        // }
-
-        // [HttpPut("{id}")]
-        // public async Task<IActionResult> Update(int id, [FromBody] Book book) {
-        //     if (book == null ) {
-        //         return BadRequest("Book data is invalid or IDs do not match.");
-        //     }
-        //     var bookUpdate = _context.Book.Find(id);
-        //     if (bookUpdate == null) {
-        //         return NotFound("Book not found.");
-        //     }
-        //     try {
-        //         bookUpdate.Title = book.Title;
-        //         bookUpdate.Description = book.Description;
-        //         bookUpdate.Price = book.Price;
-        //         bookUpdate.Amount = book.Amount;
-        //         _context.Book.Update(bookUpdate);
-        //         await _context.SaveChangesAsync();
-        //         return Ok(bookUpdate);
-        //     }
-        //     catch (Exception ex) {
-        //         return StatusCode(500, $"Internal server error: {ex.Message}");
-        //     }
-        // }
-
-        // [HttpDelete("{id}")]
-        // public async Task<IActionResult> Delete(int id) {
-        //     var book = _context.Book.FirstOrDefault(x => x.Id == id); {
-        //         if(book == null) {
-        //             return NotFound();
-        //         }
-
-        //         _context.Book.Remove(book);
-        //         await _context.SaveChangesAsync();
-        //         return NoContent();
-        //     }
-        // }
 
         [HttpGet]
         public async Task<IActionResult> GetAllBooks() {
+
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var books = await _bookRepo.GetAllAsync();
             var bookDto = books.Select(s => s.ToBookDto());
-            return Ok(books);
+            return Ok(bookDto);
         }
 
-        [HttpGet("{id}")]
-        public async Task<IActionResult> GetByIdBook(int id) {
+
+
+        [HttpGet("{id:int}")]
+        public async Task<IActionResult> GetByIdBook([FromRoute] int id) {
+
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var book = await _bookRepo.GetByIdAsync(id);
             if(book == null) {
                 return NotFound();
@@ -106,35 +47,34 @@ namespace api.controllers {
             return Ok(book.ToBookDto());
         }
 
-        [HttpPost("{customerId}")]
-        public async Task<IActionResult> CreateBook([FromRoute]int customerId, CreateBookDto bookDto) {
-            // var bookModel = bookDto.CreateBookDto();
-            // await _bookRepo.CreateAsync(bookModel);
-            // return CreatedAtAction(nameof(GetByIdBook), new {id = bookModel.Id},bookModel.ToBookDto());
-            {
-                if(!await _customerRepo.CustomerExists(customerId)) {
-                    return BadRequest("Customer Does Not Exist");
-                }
 
-                var bookModel = bookDto.CreateBookDto(customerId);
-                await _bookRepo.CreateAsync(bookModel);
-                return CreatedAtAction(nameof(GetByIdBook), new {id = bookModel}, bookModel.ToBookDto());
+
+        // [HttpPost("{customerId:int}")]
+        [HttpPost]
+        [Route("{customerId:int}")]
+
+        public async Task<IActionResult> CreateBook([FromRoute]int customerId, CreateBookDto bookDto) {
+
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
+            if(!await _customerRepo.CustomerExists(customerId)) {
+                return BadRequest("Customer Does Not Exist 3999");
             }
+
+            var bookModel = bookDto.CreateBookDto(customerId);
+            await _bookRepo.CreateAsync(bookModel);
+            return CreatedAtAction(nameof(GetByIdBook), new {id = bookModel.Id}, bookModel.ToBookDto());
         }
 
-        // [HttpPut]
-        // public async Task<IActionResult> UpdateBook(int id, UpdateBookDto bookDto) {
-        //     var bookUpdate = await _bookRepo.UpdateAsync(id, bookDto);
-
-        //     if(bookUpdate == null) {
-        //         return NotFound();
-        //     }
-        //     return Ok(bookUpdate.ToBookDto());
-        // }
 
 
-        [HttpPut("{id}")]
+        [HttpPut("{id:int}")]
         public async Task<IActionResult> UpdateBook([FromRoute] int id, UpdateBookDto bookDto) {
+            
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+                
             var bookUpdate = await _bookRepo.UpdateAsync(id, bookDto);
 
             if(bookUpdate == null) {
@@ -145,14 +85,17 @@ namespace api.controllers {
 
 
 
-        [HttpDelete("{id}")]
+        [HttpDelete("{id:int}")]
         public async Task<IActionResult> DeleteBook([FromRoute] int id) {
-            var book = await _bookRepo.DeleteAsync(id);
+            
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
 
-                if(book == null) {
-                    return NotFound();
-                }
-                return NoContent();
+            var book = await _bookRepo.DeleteAsync(id);
+            if(book == null) {
+                return NotFound();
+            }
+            return NoContent();
         }
